AMp scorer work, folders for commands, climber work
