+++ src/main/java/frc/robot/Constants.java
@@ -44,8 +44,8 @@ public final class Constants {
 
     // Amp Neo 550
     public static final int AMP_MOTOR_CAN_ID = 6;
-    public static final int AMP_MOTOR_SMART_CURRENT_LIMIT = 20; // NEO 550 
-    public static final int AMP_MOTOR_SECONDARY_CURRENT_LIMIT = 40;
+    public static final int AMP_MOTOR_SMART_CURRENT_LIMIT = 10; // NEO 550 
+    public static final int AMP_MOTOR_SECONDARY_CURRENT_LIMIT = 20;
     public static final double AMP_UP_PCT = 10.0;     // TODO: confirm these values and direction
     public static final double AMP_DOWN_PCT = -10.0;   // TODO: confirm these values and direction
     public static final double FR_SHOOTER_MOTOR_PCT_AMP = 45.0;
