+++ src/main/java/frc/robot/commands/Indexing/RunIndexUpCmd.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Indexing;
 
 import org.livoniawarriors.REVColorSensor;
