+++ src/main/java/frc/robot/commands/Indexing/RunIndexUpContinuousCmd.java
@@ -2,19 +2,27 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Indexing;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.ShooterSubSys;
+import frc.robot.subsystems.IndexerSubSys;
 
-public class PrimeShooterSpeakerCmd extends Command {
-    private final ShooterSubSys shooterSubSysObj;
+public class RunIndexUpContinuousCmd extends Command {
+    /** Creates a new RunIndexUp. */
+    private final IndexerSubSys indexerSubSysObj;
 
-    /** Creates a new ShootNoteCmd. */
-    public PrimeShooterSpeakerCmd(ShooterSubSys shooterSubSysObj) {
-        this.shooterSubSysObj = shooterSubSysObj;
-        addRequirements(shooterSubSysObj);
+    /**
+     * Creates a new RunIndexUpCmd.
+     * Makes indexerSubSysObj a requirement
+     * 
+     * @param indexerSubSysObj The IndexerSubsystem from the where it is being
+     *                         called
+     */
+
+    public RunIndexUpContinuousCmd(IndexerSubSys indexerSubSysObj) {
         // Use addRequirements() here to declare subsystem dependencies.
+        this.indexerSubSysObj = indexerSubSysObj;
+        addRequirements(indexerSubSysObj);
     }
 
     // Called when the command is initially scheduled.
@@ -26,13 +34,13 @@ public class PrimeShooterSpeakerCmd extends Command {
     // Called every time the scheduler runs while the command is scheduled.
     @Override
     public void execute() {
-        shooterSubSysObj.runShooterHighSpeed();
+        indexerSubSysObj.runIndexerUp();
     }
 
     // Called once the command ends or is interrupted.
     @Override
     public void end(boolean interrupted) {
-        shooterSubSysObj.stopShooter();
+        indexerSubSysObj.stopIndexMotors();
     }
 
     // Returns true when the command should end.
