+++ src/main/java/frc/robot/commands/ShootingAmp/LowerAmpCmd.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.ShootingAmp;
 
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.Command;
@@ -12,7 +12,7 @@ public class LowerAmpCmd extends Command {
   private final AmpScorerSubSys ampScorerSubSysObj;
   private static final Timer TIMER = new Timer();
 
-  private double timerLim = 7.5;
+  private double timerLim = 4.0;
 
   /** Creates a new LowerAmpCmd. */
   public LowerAmpCmd(AmpScorerSubSys ampScorerSubSysObj) {
