+++ src/main/java/frc/robot/commands/ShootingAmp/PrimeShooterAmpCmd.java
@@ -2,19 +2,18 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.ShootingAmp;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.ClimberSubSys;
+import frc.robot.subsystems.ShooterSubSys;
 
-public class ClimbUpCmd extends Command {
-    private final ClimberSubSys climberSubSysObj;
-
-    /** Creates a new ClimbUpCmd. */
-    public ClimbUpCmd(ClimberSubSys climberSubSysObj) {
-        this.climberSubSysObj = climberSubSysObj;
+public class PrimeShooterAmpCmd extends Command {
+    private final ShooterSubSys shooterSubSysObj;
+    /** Creates a new PrimeShooterAmpCmd. */
+    public PrimeShooterAmpCmd(ShooterSubSys shooterSubSysObj) {
+        this.shooterSubSysObj = shooterSubSysObj;
+        addRequirements(shooterSubSysObj);
         // Use addRequirements() here to declare subsystem dependencies.
-        addRequirements(climberSubSysObj);
     }
 
     // Called when the command is initially scheduled.
@@ -26,13 +25,13 @@ public class ClimbUpCmd extends Command {
     // Called every time the scheduler runs while the command is scheduled.
     @Override
     public void execute() {
-        climberSubSysObj.runClimberUp();
+        shooterSubSysObj.runShooterLowSpeed();
     }
 
     // Called once the command ends or is interrupted.
     @Override
     public void end(boolean interrupted) {
-        climberSubSysObj.stopClimberMotor();
+        shooterSubSysObj.stopShooter();
     }
 
     // Returns true when the command should end.
