+++ src/main/java/frc/robot/commands/ShootingSpeaker/AngleShooterDown.java
@@ -2,42 +2,46 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.ShootingSpeaker;
 
+import edu.wpi.first.units.Angle;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.ShooterAnglerSubSys;
 
-public class AngleShooterUp extends Command {
+public class AngleShooterDown extends Command {
     private final ShooterAnglerSubSys shooterAnglerSubSysObj;
     private static final Timer TIMER = new Timer();
-    private final double timerLim;
+    private double timerLim = 0.6;
 
-    /** Creates a new AngleShooterUp. */
-    public AngleShooterUp(ShooterAnglerSubSys shooterAnglerSubSysObj) {
+    /** Creates a new AngleShooterDown. */
+    public AngleShooterDown(ShooterAnglerSubSys shooterAnglerSubSysObj) {
         this.shooterAnglerSubSysObj = shooterAnglerSubSysObj;
-        timerLim = 3.0;
-        // Use addRequirements() here to declare subsystem dependencies.
         addRequirements(shooterAnglerSubSysObj);
+        // Use addRequirements() here to declare subsystem dependencies.
     }
 
     // Called when the command is initially scheduled.
     @Override
     public void initialize() {
         TIMER.restart();
-        // No op
+    
     }
 
     // Called every time the scheduler runs while the command is scheduled.
     @Override
     public void execute() {
-        shooterAnglerSubSysObj.runLinearActuator();
+        while (TIMER.get() < timerLim) {
+            shooterAnglerSubSysObj.runLinearActuatorReverse(); //angle shooter down to about 42 deg.
+        }
+        shooterAnglerSubSysObj.stopLinearActuator();
     }
 
     // Called once the command ends or is interrupted.
     @Override
     public void end(boolean interrupted) {
-        shooterAnglerSubSysObj.stopLinearActuator();
+        // shooterAnglerSubSysObj.stopLinearActuator();   
+        shooterAnglerSubSysObj.runLinearActuator();  // angle shooter back to 60 deg.
         TIMER.stop();
         TIMER.reset();
     }
@@ -45,6 +49,6 @@ public class AngleShooterUp extends Command {
     // Returns true when the command should end.
     @Override
     public boolean isFinished() {
-        return TIMER.get() > timerLim;
+        return false;
     }
 }
