+++ src/main/java/frc/robot/commands/ShootingSpeaker/AngleShooterUp.java
@@ -2,46 +2,42 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.ShootingSpeaker;
 
-import edu.wpi.first.units.Angle;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.ShooterAnglerSubSys;
 
-public class AngleShooterDown extends Command {
+public class AngleShooterUp extends Command {
     private final ShooterAnglerSubSys shooterAnglerSubSysObj;
     private static final Timer TIMER = new Timer();
-    private double timerLim = 0.6;
+    private final double timerLim;
 
-    /** Creates a new AngleShooterDown. */
-    public AngleShooterDown(ShooterAnglerSubSys shooterAnglerSubSysObj) {
+    /** Creates a new AngleShooterUp. */
+    public AngleShooterUp(ShooterAnglerSubSys shooterAnglerSubSysObj) {
         this.shooterAnglerSubSysObj = shooterAnglerSubSysObj;
-        addRequirements(shooterAnglerSubSysObj);
+        timerLim = 3.0;
         // Use addRequirements() here to declare subsystem dependencies.
+        addRequirements(shooterAnglerSubSysObj);
     }
 
     // Called when the command is initially scheduled.
     @Override
     public void initialize() {
         TIMER.restart();
-    
+        // No op
     }
 
     // Called every time the scheduler runs while the command is scheduled.
     @Override
     public void execute() {
-        while (TIMER.get() < timerLim) {
-            shooterAnglerSubSysObj.runLinearActuatorReverse();
-        }
-        shooterAnglerSubSysObj.stopLinearActuator();
+        shooterAnglerSubSysObj.runLinearActuator();
     }
 
     // Called once the command ends or is interrupted.
     @Override
     public void end(boolean interrupted) {
-        // shooterAnglerSubSysObj.stopLinearActuator();   
-        shooterAnglerSubSysObj.runLinearActuator();
+        shooterAnglerSubSysObj.stopLinearActuator();
         TIMER.stop();
         TIMER.reset();
     }
@@ -49,6 +45,6 @@ public class AngleShooterDown extends Command {
     // Returns true when the command should end.
     @Override
     public boolean isFinished() {
-        return false;
+        return TIMER.get() > timerLim;
     }
 }
