+++ src/main/java/frc/robot/commands/ShootingSpeaker/PrimeShooterSpeakerCmd.java
@@ -2,18 +2,18 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.ShootingSpeaker;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.IndexerSubSys;
+import frc.robot.subsystems.ShooterSubSys;
 
-public class RunIndexDownCmd extends Command {
-    private final IndexerSubSys indexerSubSysObj;
+public class PrimeShooterSpeakerCmd extends Command {
+    private final ShooterSubSys shooterSubSysObj;
 
-    /** Creates a new RunIndexDown. */
-    public RunIndexDownCmd(IndexerSubSys indexerSubSys) {
-        this.indexerSubSysObj = indexerSubSys;
-        addRequirements(indexerSubSys);
+    /** Creates a new ShootNoteCmd. */
+    public PrimeShooterSpeakerCmd(ShooterSubSys shooterSubSysObj) {
+        this.shooterSubSysObj = shooterSubSysObj;
+        addRequirements(shooterSubSysObj);
         // Use addRequirements() here to declare subsystem dependencies.
     }
 
@@ -26,13 +26,13 @@ public class RunIndexDownCmd extends Command {
     // Called every time the scheduler runs while the command is scheduled.
     @Override
     public void execute() {
-        indexerSubSysObj.runIndexerDown();
+        shooterSubSysObj.runShooterHighSpeed();
     }
 
     // Called once the command ends or is interrupted.
     @Override
     public void end(boolean interrupted) {
-        indexerSubSysObj.stopIndexMotors();
+        shooterSubSysObj.stopShooter();
     }
 
     // Returns true when the command should end.
