+++ src/main/java/org/livoniawarriors/odometry/PigeonGyro.java
@@ -36,6 +36,8 @@ public class PigeonGyro implements IGyroHardware {
         return yprDeg[0];  //was negative when Pigeon was upside down.
     }
 
+
+    @Override
     public double zeroGyroAngle() {
         return yprDeg[0] = 0.0; 
     }
